unittest numpy fix #203

+++ python/test/unittest_setup_teardown.py
@@ -58,7 +58,7 @@ class DummyUnitTest(unittest.TestCase):
         pd.testing.assert_frame_equal(self.df, df_target)  # same as before
 
     def test_numpy(self):
-        np_target = np.array([1, 2, 3], dtype=np.float32)
+        np_target = np.array([1, 2, 3])
         np_target2 = np.array([0.9999, 2, 3])
         self.assertTrue((self.np_array == np_target).all())
         np.testing.assert_array_equal(self.np_array, np_target)
